pyproject.toml
[project]
name = "discord-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncio>=4.0.0",
    "cachetools>=6.2.1",
    "discord-py>=2.7.1",
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic>=2.12.3",
    "pymongo>=4.15.3",
]
